Overhaul of list handling in both Array and Hash
This commit updates Array and Hash to return a Collection::Array or
Collection::Hash object as appropriate whenever a list was being returned. This
allows for chaining of method calls such as $foo->map()->grep()->elements;

To support expected behavior similar to Moose::NativeTraits, Array::AutoFlatten
and Hash::AutoFlatten classes are provided. This will be what MooX::HandlesVia
will consume.